[modules] Moved edgeflow.c and .h to lib/vision folder
