+++ conf/modules/cv_opticflow.xml
@@ -90,11 +90,10 @@
     <file name="pprz_algebra_float.c" dir="math"/>
     <file name="pprz_matrix_decomp_float.c" dir="math"/>
 
-    <file name="edge_flow.c" dir="modules/computer_vision/opticflow"/>
-
     <!-- Main vision calculations -->
     <file name="fast_rosten.c" dir="modules/computer_vision/lib/vision"/>
     <file name="lucas_kanade.c" dir="modules/computer_vision/lib/vision"/>
+    <file name="edge_flow.c" dir="modules/computer_vision/lib/vision"/>
 
   <raw>
       include $(CFG_SHARED)/udp.makefile
